+++ src/lib.rs
@@ -4,7 +4,7 @@ pub mod model;
 pub mod macros;
 pub mod dto;
 pub mod middleware;
-pub mod file_writer;
+pub mod io;
 
 // #[cfg(test)] #[macro_use]
 // extern crate assert_matches;
